+++ bin/Builder.php
@@ -4,50 +4,27 @@ declare(strict_types=1);
 
 namespace Rowbot\Idna\Bin;
 
+use Rowbot\Idna\Idna;
 use RuntimeException;
 
 use function array_map;
 use function explode;
 use function fclose;
-use function file_get_contents;
 use function fopen;
 use function intval;
-use function json_decode;
-use function json_last_error;
-use function json_last_error_msg;
 use function sprintf;
 use function usort;
-
-use const DIRECTORY_SEPARATOR as DS;
-use const JSON_ERROR_NONE;
 
 abstract class Builder
 {
     protected const BASE_URL = 'https://www.unicode.org/Public';
-
-    protected static function getUnicodeVersion(): string
-    {
-        $data = file_get_contents(__DIR__ . DS . '..' . DS . 'composer.json');
-
-        if ($data === false) {
-            throw new RuntimeException('Failed to read composer.json.');
-        }
-
-        $composerConfig = json_decode($data);
-
-        if ($composerConfig === null || json_last_error() !== JSON_ERROR_NONE) {
-            throw new RuntimeException(sprintf('JSON decoding composer.json failed: %s', json_last_error_msg()));
-        }
-
-        return $composerConfig->extra->idna->unicode_version;
-    }
 
     /**
      * @return resource
      */
     protected static function getIdnaDataResource(string $file)
     {
-        $file = sprintf('%s/idna/%s/%s', self::BASE_URL, self::getUnicodeVersion(), $file);
+        $file = sprintf('%s/idna/%s/%s', self::BASE_URL, Idna::UNICODE_VERSION, $file);
         $handle = fopen($file, 'r');
 
         if ($handle === false) {
@@ -62,7 +39,7 @@ abstract class Builder
      */
     protected static function getUnicodeDataResource(string $file)
     {
-        $file = sprintf('%s/%s/ucd/%s', self::BASE_URL, self::getUnicodeVersion(), $file);
+        $file = sprintf('%s/%s/ucd/%s', self::BASE_URL, Idna::UNICODE_VERSION, $file);
         $handle = fopen($file, 'r');
 
         if ($handle === false) {