+++ src/Idna.php
@@ -22,6 +22,8 @@ use function substr;
  */
 final class Idna
 {
+    public const UNICODE_VERSION = '14.0.0';
+
     private const DEFAULT_UNICODE_OPTIONS = [
         'CheckHyphens'            => true,
         'CheckBidi'               => true,